+++ src/main/java/com/msd/erp/application/views/PurchaseOrderLineDTO.java
@@ -3,7 +3,6 @@ package com.msd.erp.application.views;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
@@ -12,7 +11,7 @@ public class PurchaseOrderLineDTO {
     private Long purchaseOrderId;
     private Long articleId;
     private Integer quantity;
-    private BigDecimal totalLineAmount;
-    private BigDecimal totalLineAmountWithVAT;
-    private BigDecimal price;
+    private Double totalLineAmount;
+    private Double totalLineAmountWithVAT;
+    private Double price;
 }
